+++ include/fase1/validacao.h
@@ -2,7 +2,7 @@
 #define VALIDACAO_H
 
 #include <stdio.h>
-#include "../include/carregamento.h"
+#include "carregamento.h"
 
 void validar_arquivo(Produto *vetor, int linhas); //função para validar entradas
