+++ testes/fase1/teste_busca.c
