+++ testes/fase1/teste_carregamento.c
