+++ testes/fase1/teste_temporizador.c
